# MM-57344: Use a row constructor comparison to avoid a filter