@mattermost-build 2024-03-19
Successfully triggered E2E testing!
[GitLab pipeline](http://git.internal.mattermost.com/qa/cypress-ui-automation/-/pipelines/414948) | [Test dashboard](https://automation-dashboard.vercel.app/cycle/414948)