@agarciamontoro 2024-03-19
@saturninoabril, I think the E2E tests look good, but could you take a look as well? That would give me some peace of mind.