@saturninoabril 2024-03-19
Yes, that looks good, thanks for checking with me @agarciamontoro.