@streamer45 2024-03-20
Thank you for this effort @agarciamontoro! :100: 
Would it make sense to run a sanity check on MySQL before merging?