@agarciamontoro 2024-03-20
>  Would it make sense to run a sanity check on MySQL before merging?

Yes, I'm spinning up a deployment right now to check